Move headers into folders
