+++ Sources/COpenCombineHelpers/Locking.cpp
