+++ Sources/COpenCombineHelpers/swift/ABI/MetadataKind.def
