+++ Sources/COpenCombineHelpers/swift/ABI/TrailingObjects.h
