+++ Sources/COpenCombineHelpers/swift/Basic/FlagSet.h
